heavy refactor for tn-test-utils and test-utils feature flags
