+++ crates/consensus/primary/src/test_utils/committee.rs
@@ -1,11 +1,10 @@
 //! Committe fixture for all authorities and their workers within a committee for a specific epoch.
 
 use super::{AuthorityFixture, Builder};
-use crate::fixture_batch_with_transactions;
 use std::collections::{BTreeMap, BTreeSet};
 use tn_types::{
-    AuthorityIdentifier, Certificate, CertificateDigest, Committee, Database, Hash as _, Header,
-    HeaderBuilder, Round, Vote, WorkerCache,
+    fixture_batch_with_transactions, AuthorityIdentifier, Certificate, CertificateDigest,
+    Committee, Database, Hash as _, Header, HeaderBuilder, Round, Vote, WorkerCache,
 };
 
 /// Fixture representing a committee to reach consensus.
