+++ crates/consensus/primary/src/test_utils/mod.rs
@@ -1,0 +1,8 @@
+//! Test utils for authorities.
+
+mod authority;
+mod builder;
+mod committee;
+pub use authority::*;
+pub use builder::*;
+pub use committee::*;
