+++ crates/consensus/primary/src/tests/certificate_tests.rs
@@ -5,8 +5,8 @@ use rand::{
     SeedableRng,
 };
 use std::num::NonZeroUsize;
+use tn_primary::CommitteeFixture;
 use tn_storage::mem_db::MemDatabase;
-use tn_test_utils::CommitteeFixture;
 use tn_types::{AuthorityIdentifier, BlsKeypair, Certificate, SignatureVerificationState, Vote};
 
 #[test]
