+++ crates/consensus/primary/src/tests/proposer_tests.rs
@@ -2,8 +2,8 @@
 
 use super::*;
 use crate::consensus::LeaderSwapTable;
+use tn_primary::CommitteeFixture;
 use tn_storage::mem_db::MemDatabase;
-use tn_test_utils::CommitteeFixture;
 use tn_types::B256;
 
 #[tokio::test]
