+++ crates/consensus/primary/tests/certificate_order_test.rs
@@ -3,8 +3,8 @@
 use indexmap::IndexMap;
 use rand::{rngs::OsRng, seq::SliceRandom};
 use std::{collections::BTreeSet, num::NonZeroUsize};
+use tn_primary::{AuthorityFixture, CommitteeFixture};
 use tn_storage::mem_db::MemDatabase;
-use tn_test_utils::{AuthorityFixture, CommitteeFixture};
 use tn_types::{
     AuthorityIdentifier, BlockNumHash, BlsPublicKey, BlsSignature, Certificate, Committee, Header,
     Vote, VotingPower,
