+++ crates/test-utils/src/lib.rs
@@ -1,31 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
-mod authority;
-pub use authority::*;
 mod execution;
 pub use execution::*;
-mod worker;
-pub use worker::*;
-
-pub mod committee;
-pub use committee::*;
-
-pub mod builder;
-pub use builder::*;
-
-pub use execution::{
-    default_test_execution_node, execution_builder, faucet_test_execution_node, CommandParser,
-    TestExecutionNode,
-};
-
-pub mod helpers;
-pub use helpers::*;
-
 mod temp_dirs;
 pub use temp_dirs::*;
-
-mod tracing;
-pub use tracing::init_test_tracing;
 
 #[cfg(test)]
 #[path = "tests/output_tests.rs"]
